requirements.txt
discord.py==1.7.3
requests~=2.28.1
gspread~=5.4.0
openai~=0.22.1